Fixed hunspell compiling by MinGW 64

+++ hunspell/csutil.cxx
@@ -217,7 +217,7 @@ int flag_bsearch(unsigned short flags[], unsigned short flag, int length) {
       }
       if (dp) {
          *stringp = dp+1;
-         int nc = (int)((unsigned long)dp - (unsigned long)mp);
+         ptrdiff_t nc = dp - mp;
          *(mp+nc) = '\0';
       } else {
          *stringp = mp + strlen(mp);
